+++ lib/src/async/execute_driver.dart
@@ -10,11 +10,11 @@ class ExecuteDriver {
 
   ExecuteDriver(this._client, this._handler);
 
-  Future<Map<String, dynamic>> executeDriver(
-          String script, String type, Duration timeoutMs) =>
+  Future<Map<String, dynamic>> executeDriver(String script,
+          {String type, Duration timeout}) =>
       _client.send(
           _handler.executeDriver
-              .buildExecuteDriverRequest(script, type, timeoutMs),
+              .buildExecuteDriverRequest(script, type: type, timeout: timeout),
           _handler.executeDriver.parseExecuteDriverResponse);
 
   @override